created model_interactor.py
