+++ backend/data_formatter.py
@@ -144,19 +144,7 @@ class DataFormatter:
 
         return display_format
 
-    @staticmethod
-    def _add_back_missing(df: pd.DataFrame) -> pd.DataFrame:
-        df.set_index('date', inplace=True)
-        all_dates = pd.date_range(start=df.index.min(), end=df.index.max(), freq='D')
-        df = DataFormatter._reindex_with_dates(all_dates, df)
-        return df
 
-    @staticmethod
-    def _reindex_with_dates(all_dates, df):
-        df = df.reindex(all_dates, fill_value=0)
-        df.index = df.index.date
-        df.index.name = 'date'
-        return df
 
     def get_for_predicting(self) -> tuple[pd.DataFrame, pd.DataFrame]:
         """Formats the data for out."""
